+++ ui/CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the **Prowler UI** are documented in this file.
 - RSS feeds support [(#9109)](https://github.com/prowler-cloud/prowler/pull/9109)
 - Customer Support menu item [(#9143)](https://github.com/prowler-cloud/prowler/pull/9143)
 - PDF reporting for ENS compliance framework [(#9158)](https://github.com/prowler-cloud/prowler/pull/9158)
+- IaC (Infrastructure as Code) provider support for scanning remote repositories [(#8751)](https://github.com/prowler-cloud/prowler/pull/8751)
 
 ### 🔄 Changed
 
@@ -21,7 +22,6 @@ All notable changes to the **Prowler UI** are documented in this file.
 ### 🔄 Changed
 
 - Upgrade React to version 19.2.0 [(#9039)](https://github.com/prowler-cloud/prowler/pull/9039)
-- Add support for Oracle Cloud Infrastructure (OCI) provider [(#8984)](https://github.com/prowler-cloud/prowler/pull/8984)
 
 ---
 
@@ -87,6 +87,8 @@ All notable changes to the **Prowler UI** are documented in this file.
 
 - Field-level email validation message [(#8698)](https://github.com/prowler-cloud/prowler/pull/8698)
 - POST method on auth form [(#8699)](https://github.com/prowler-cloud/prowler/pull/8699)
+
+---
 
 ## [1.12.0] (Prowler v5.12.0)